+++ instrumentation/opentelemetry-instrumentation-redis/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Update default SpanKind to `SpanKind.CLIENT` ([#965](https://github.com/open-telemetry/opentelemetry-python/pull/965))
 - Change package name to opentelemetry-instrumentation-redis
   ([#966](https://github.com/open-telemetry/opentelemetry-python/pull/966))